+++ src/config/languages/englishContent.js
@@ -19,12 +19,12 @@ const englishContent = {
         banner: {
             title: 'Hola world, I am Gabi Polanco-Ferreyra',
             subtitle:
-                'My background in law allowed me to develop my empathic view of the world, and my limitless curiosity led me to study different design programs in a self-taught and formal way.',
-            subtitle2: 'The result: Gabi,  ',
-            subtitle3: 'new UX/UI designer who wants to be part of the ',
+                'My background in law allowed me to develop and exercise my empathic view of the world; my limitless curiosity and desire to create led me to study design in a self-taught and formal way.',
+            subtitle2: 'The result is: Gabi, a new UX/UI',
+            subtitle3: 'designer who wants to be part of the ',
             subtitle4: "I'm driven by the desire to make this world more accessible, fair, and democratic.",
             subtitleBold1: '(very empathic)',
-            subtitleBold2: 'digital revolution'
+            subtitleBold2: 'digital revolution.'
         },
         selectedWorks: {
             title: 'Selected works',
@@ -34,14 +34,14 @@ const englishContent = {
         about: {
             title: 'about me',
             content:
-                'Soy Gabi Polanco-Ferreyra and I come from a colorful background, graduated from Law School, did a master’s in Mediation and Conflict Resolutions, worked in government for many years before I dedicated myself to being a full-time traveler around the world for a few months. I tried a lot of other jobs before committing to my creative talents full-time.I love to travel, see the world, camp, hike mountains, and meet people (really). My dream is to sail around the world one day.',
+                'I am Gabi Polanco-Ferreyra and I come from a colorful background. I graduated from Law School, did a master’s in Mediation and Conflict Resolutions, and worked in government for many years before I dedicated myself to being a full-time traveler around the world for a few months. I tried a lot of other jobs before committing to my creative talents full-time. I love to travel, see the world, camp, hike mountains, and meet people (really). My dream is to sail around the world one day.',
             contentHighlight: 'I am on my creative-human journey, and ME ENCANTA.',
             linkedinLinkText: 'Linkedin',
             linkedinLink: 'https://www.linkedin.com/in/gabriela-polanco-ferreyra-b5b25717a/',
             mediumLinkText: 'Medium',
             mediumLink: 'https://medium.com/@gabrielapolancoferreyra',
-            cvLinkText: 'CV',
-            cvLink: '/resume.pdf',
+            cvLinkText: 'Resume',
+            cvLink: '/Gabriela Polanco-Ferreyra Resume.pdf',
             behanceLinkText: 'Behance',
             behanceLink: 'https://www.behance.net/gabrielapolanco2',
             askForAvailability: "let's talk"