+++ src/data/webDesign.js
@@ -4,10 +4,10 @@ const array = [
         website: 'https://tierra-roja.herokuapp.com/',
         title: 'Tierra roja',
         en: {
-            subtitle: 'Web app for clay artists in Misiones.'
+            subtitle: 'Web app for clay artists in Misiones'
         },
         es: {
-            subtitle: 'Web app for clay artists in Misiones.'
+            subtitle: 'Web app for clay artists in Misiones'
         },
         type: 'web-design',
         status: 'ready'
@@ -17,10 +17,10 @@ const array = [
         website: 'https://www.chapati-delta.com.ar/',
         title: 'Chapati',
         en: {
-            subtitle: "Conscious website promoting the Chapati' project"
+            subtitle: "Conscious website promoting the Chapati's project"
         },
         es: {
-            subtitle: "Conscious website promoting the Chapati' project"
+            subtitle: "Conscious website promoting the Chapati's project"
         },
         type: 'web-design',
         status: 'ready'
@@ -30,10 +30,10 @@ const array = [
         title: 'Tomi Scattini',
         website: 'https://tomiscattini.com/',
         en: {
-            subtitle: 'Full stack web developer Portfolio'
+            subtitle: 'Full stack developer portfolio'
         },
         es: {
-            subtitle: 'Full stack web developer Portfolio'
+            subtitle: 'Full stack developer portfolio'
         },
         type: 'web-design',
         status: 'ready'
@@ -43,7 +43,7 @@ const array = [
         website: 'https://vaikuntha.com.ar/',
         title: 'Vaikuntha',
         en: {
-            subtitle: 'Ceramist Portfolio (coming soon)'
+            subtitle: 'Ceramist portfolio (coming soon)'
         },
         es: {
             subtitle: 'Portfolio de Ceramista (próximanente)'